iter11 Меняем тип у ID ссылок, рефакторим логику открытия файла для ссылок

+++ internal/models/models.go
@@ -15,7 +15,7 @@ type (
 	}
 
 	Shorten struct {
-		UUID        int8   `json:"uuid"`
+		UUID        int64  `json:"uuid"`
 		ShortURL    string `json:"short_url"`
 		OriginalURL string `json:"original_url"`
 	}
